pyproject.toml
[project]
name = "drf_react_ecommerce"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "django>=6.0",
    "django-autoslug>=1.9.9",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "djangorestframework-stubs>=3.16.6",
    "drf-spectacular>=0.29.0",
    "pillow>=12.1.0",
]
